#2 - kbhit 헤더파일로 분리
