+++ kbhit.h
@@ -1,0 +1,9 @@
+#ifndef _KBHIT_
+#define _KBHIT_
+
+#include <ncurses.h>
+#include <iostream>
+
+int kbhit(void);
+
+#endif
